shibuya: Modify the GPIO69 to SPI mode in kernel
As the SPI3 CS pin, GPIO69 is pulled high during the lk phase,
which causes a voltage step when IOVCC is powered on.
So change the gpio69 to gpio mode in lk,
and change the gpio69 to spi mode in kenrel
This step issue can be resolved by switching GPIO69 to GPIO mode.

Change-Id: Ieaa212b76d16a4f5aa86393b3c8b1fce84aa5612
Reviewed-on: https://gerrit.mot.com/3642888
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Obi-Wan Kenobi
Reviewed-by: <lvxj4@motorola.com>
Reviewed-by: Yeqing Wang <wangyq13@motorola.com>
Reviewed-by: Hongshu Lou <louhs1@motorola.com>
Submit-Approved: Jira Key